producer_service/requirements.txt
django==4.2
djangorestframework==3.14.0
requests==2.31.0
